Fixes for TravisCI testing. - Run g++ non-optimised compilation as second job, since it's likely to be the second-longest job. - Prevent g++ compilation error due to failure to inline a function, where that failure occurs due to that function being recursive and the compiler has been instructed to not optimise (dirgen appears to have been the only instance of this occurring). - Fix comparison between signed and unsigned integers in DICOM code.

+++ core/file/dicom/mapper.cpp
@@ -215,7 +215,7 @@ namespace MR {
           H.size(1) = std::floor (frame.dim[1] / mosaic_size);
           H.size(2) = image.images_in_mosaic;
 
-          if (frame.acq_dim[0] > H.size(0) || frame.acq_dim[1] > H.size(1)) {
+          if (frame.acq_dim[0] > size_t(H.size(0)) || frame.acq_dim[1] > size_t(H.size(1))) {
             WARN ("acquisition matrix [ " + str (frame.acq_dim[0]) + " " + str (frame.acq_dim[1])
                 + " ] is smaller than expected [ " + str(H.size(0)) + " " + str(H.size(1)) + " ] in DICOM mosaic");
             WARN ("  image may be incorrectly reformatted");
@@ -228,7 +228,7 @@ namespace MR {
             WARN ("  image may be incorrectly reformatted");
           }
 
-          if (frame.acq_dim[0] != H.size(0)|| frame.acq_dim[1] != H.size(1))
+          if (frame.acq_dim[0] != size_t(H.size(0)) || frame.acq_dim[1] != size_t(H.size(1)))
             INFO ("note: acquisition matrix [ " + str (frame.acq_dim[0]) + " " + str (frame.acq_dim[1])
                 + " ] differs from reconstructed matrix [ " + str(H.size(0)) + " " + str(H.size(1)) + " ]");
